move getting genesis artifacts to own file
