+++ tools/config/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod config_cli;
 pub mod config_wizard;
+pub mod get_genesis_artifacts;
 pub mod make_profile; // TODO: deprecated?
 pub mod make_yaml_public_fullnode;
 pub mod make_yaml_validator;
