+++ tools/storage/src/lib.rs
@@ -1,8 +1,8 @@
+pub mod bootstrap;
 pub mod dbtool_init;
 pub mod download_bundle;
-mod parse_folder_names;
+pub mod parse_folder_names;
 pub mod read_snapshot;
-pub mod read_tx_chunk;
 pub mod restore;
 pub mod restore_bundle;
 pub mod storage_cli;
